ManualDataValidator/WebDataTracker/WebDataTracker/requirements.txt
flask
pymongo
python-dotenv
openai
python-levenshtein
